+++ src/easter/Est_dir_util.c
@@ -288,7 +288,7 @@ isOldFile( const char* file_path, void* arg )
 	
 		if( ZnkDate_compareDay( &date, &threshold ) >= 0 ){
 			diff_sec = ZnkDate_diffSecond( &current, &date, 1 );
-			if( diff_sec >= sec_ago ){
+			if( diff_sec >= (long)sec_ago ){
 				is_old = true;
 			}
 		} else {
